+++ src/index.ts
@@ -4,7 +4,7 @@ export * from "./reactive/index";
 export * from "./utils/index";
 export { genUID } from "./helper/generation";
 export { randInt } from "./helper/random";
-export { mounterNode } from "./mounter/index";
+export { mounterComponent } from "./mounter/index";
 export * from "./types";
 export * from "./jsx";
 export * from "./utils/composables";
