+++ src/mounter/dom/patch/class.ts
