+++ src/mounter/dom/patch/image.ts
